HL7Utils moved into /java/ package from /bean/common/util/ package
